utils.py cleanup.
I just applied a couple principles:

- This should not be an API of Errbot
- If not used in Errbot, it is removed
- If is it used only once in Errbot, it is moved to the spot where it is
used
- If it is used a couple times in Errbot and doesn't have a natural spot
within the source tree, it stays.
