+++ tests/utils_test.py
@@ -44,12 +44,6 @@ def test_formattimedelta():
     assert '1 hours and 13 minutes' == format_timedelta(td)
 
 
-def test_drawbar():
-    assert drawbar(5, 10) == '[████████▒▒▒▒▒▒▒]'
-    assert drawbar(0, 10) == '[▒▒▒▒▒▒▒▒▒▒▒▒▒▒▒]'
-    assert drawbar(10, 10) == '[███████████████]'
-
-
 def unescape_test():
     assert unescape_xml('&#32;') == ' '
 
@@ -74,55 +68,6 @@ def test_storage():
     assert key not in persistent_object
     assert len(persistent_object) == 0
 
-
-def test_recurse_check_structure_valid():
-    sample = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    to_check = dict(string="Foobar", list=["Foo", "Bar", "Bas"], dict={'foo': "Bar"}, none=None, true=True,
-                    false=False)
-    recurse_check_structure(sample, to_check)
-
-
-def test_recurse_check_structure_missingitem():
-    sample = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    to_check = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True)
-    with pytest.raises(ValidationException):
-        recurse_check_structure(sample, to_check)
-
-
-def test_recurse_check_structure_extrasubitem():
-    sample = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    to_check = dict(string="Foobar", list=["Foo", "Bar", "Bas"], dict={'foo': "Bar", 'Bar': "Foo"}, none=None,
-                    true=True, false=False)
-    with pytest.raises(ValidationException):
-        recurse_check_structure(sample, to_check)
-
-
-def test_recurse_check_structure_missingsubitem():
-    sample = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    to_check = dict(string="Foobar", list=["Foo", "Bar", "Bas"], dict={}, none=None, true=True, false=False)
-    with pytest.raises(ValidationException):
-        recurse_check_structure(sample, to_check)
-
-
-def test_recurse_check_structure_wrongtype_1():
-    sample = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    to_check = dict(string=None, list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    with pytest.raises(ValidationException):
-        recurse_check_structure(sample, to_check)
-
-
-def test_recurse_check_structure_wrongtype_2():
-    sample = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    to_check = dict(string="Foobar", list={'foo': "Bar"}, dict={'foo': "Bar"}, none=None, true=True, false=False)
-    with pytest.raises(ValidationException):
-        recurse_check_structure(sample, to_check)
-
-
-def test_recurse_check_structure_wrongtype_3():
-    sample = dict(string="Foobar", list=["Foo", "Bar"], dict={'foo': "Bar"}, none=None, true=True, false=False)
-    to_check = dict(string="Foobar", list=["Foo", "Bar"], dict=["Foo", "Bar"], none=None, true=True, false=False)
-    with pytest.raises(ValidationException):
-        recurse_check_structure(sample, to_check)
 
 
 def test_split_string_after_returns_original_string_when_chunksize_equals_string_size():
